+++ HISTORY.md
@@ -1,6 +1,6 @@
 # History
 
-## not yet released, version 3.13.2
+## 2017-05-26, version 3.13.2
 
 - Disabled function `chain` inside the expression parser for security
   reasons (it's not needed there anyway).